% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colnames.R
\name{add_colnames}
\alias{add_colnames}
\title{Add content to spreadsheets}
\usage{
add_colnames(names, max_row, name_num = 5)
}
\arguments{
\item{names}{column names}

\item{max_row}{cell row number}

\item{name_num}{the number of name columns to add}
}
\description{
Add column names to spreadsheet. Example: add_colnames 'as_ip_names, 3' will add first 5 vars to row 2, rest to row 3
}
