% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stuff.R
\name{add_stuff}
\alias{add_stuff}
\title{Add content to spreadsheets}
\usage{
add_stuff(sheet, row_num, colIndex, value, style)
}
\arguments{
\item{sheet}{sheet name}

\item{row_num}{cell row number}

\item{colIndex}{cell column number}

\item{value}{value to insert}

\item{style}{formatting style}
}
\description{
Used to add content to spreadsheet rows. A slightly edited version of Bert's old function.
}
