% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_dupes.R
\name{catch_dupes}
\alias{catch_dupes}
\title{Catch duplicates within a pipe.}
\usage{
catch_dupes(data, var = "provider_id", var2 = NULL, var3 = NULL)
}
\arguments{
\item{var}{by default, is provider_id}

\item{var2}{a 2nd grouping variable, like measure}

\item{var3}{3rd grouping variable, like quarter}
}
\description{
Catch duplicates within a pipe.
}
