% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix}
\usage{
cor_matrix(x)
}
\arguments{
\item{x}{data frame}
}
\description{
Quick function for formatted correlation matrix that can be saved as a data frame
}
