% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodge_legend.R
\name{dodge_legend}
\alias{dodge_legend}
\title{Move legend to avoid overlapping data points}
\usage{
dodge_legend(x, m, limit, points)
}
\arguments{
\item{x}{plot data}

\item{m}{the measure included in the plot i.e. OP-2}

\item{limit}{the highest data value that won't move the legend}

\item{points}{the number of points on the x axis. Currently includes options for 2-4.}
}
\description{
Used as argument to legend.position to move the legend to bottom when high data values would overlap it.
}
