% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_table.R
\name{double_table}
\alias{double_table}
\title{Create 2x Tables}
\usage{
double_table(data)
}
\arguments{
\item{data}{data created using FMT::reporting_rank}
}
\description{
#This function creates the side-by-side tables used in the annual reports. Access them by subsetting the list
}
