% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_just.R
\name{easy_just}
\alias{easy_just}
\title{Place Figure Text}
\usage{
easy_just(
  data,
  axis_max = 100,
  axis_min = 0,
  invert = TRUE,
  fixed = TRUE,
  upper = 1
)
}
\arguments{
\item{data}{data frame}

\item{axis_max}{maximum of fixed Y axis; default is 100; only used if fixed=T}

\item{axis_min}{minimum of fixed Y axis; default is 100; only used if fixed=T}

\item{invert}{use this option if plot shows the data in the opposite order as intended}

\item{upper}{the percentage of the Y axis that will cut off data text if exceeded; defaults to .9}
}
\description{
Used for positioning trend figure data value text appropriately in figures. A simplified version of vjust_right, which doesn't consider whether figure text may be cut off.
}
