% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_just2.R
\name{easy_just2}
\alias{easy_just2}
\title{Place Figure Text}
\usage{
easy_just2(data, axis_max = 100, axis_min = 0, invert = FALSE)
}
\arguments{
\item{data}{data frame}

\item{axis_max}{maximum of fixed Y axis; default is 100; only used if fixed=T}

\item{axis_min}{minimum of fixed Y axis; default is 100; only used if fixed=T}

\item{invert}{use this option if plot shows the data in the opposite order as intended}

\item{fixed}{whether the plot's Y axis is fixed; default is FALSE}
}
\description{
Used for positioning trend figure data value text appropriately in figures. This takes the state line into account for CAH figures.
}
