% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\docType{data}
\name{fmt_folder}
\alias{fmt_folder}
\title{FMT Folder Path}
\format{
An object of class \code{character} of length 1.
}
\usage{
fmt_folder
}
\description{
Commonly used default colors
}
\keyword{datasets}
