% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_rate.R
\name{get_report_rate}
\alias{get_report_rate}
\title{Get domain reporting rates}
\usage{
get_report_rate(data, domain)
}
\arguments{
\item{data}{data file}

\item{domain}{either IP", "OP", "HCAHPS", or "EDTC"}
}
\description{
Used in the Annual Reports to get reporting rates for each state at the state, nat, peer, and region levels
}
