% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarian.R
\name{librarian}
\alias{librarian}
\title{Load Common Packages}
\usage{
librarian(x = "basic")
}
\arguments{
\item{x}{use case: basic, stats, reports}
}
\description{
Load common packages for basic use "basic", statistics "stats", and FMT reports "reports".
}
