% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_back.R
\name{move_back}
\alias{move_back}
\title{Adjust quarters}
\usage{
move_back(period, move)
}
\arguments{
\item{period}{period, i.e. 2021Q1}

\item{move}{the number of periods back from current}
}
\description{
Used to adjust quarters/periods for MOU lists. Very simple
}
