% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edtc.R
\name{read_edtc}
\alias{read_edtc}
\title{Read EDTC data}
\usage{
read_edtc(folder, quarter, year = 2019, pivoted = FALSE)
}
\arguments{
\item{folder}{partial file path}

\item{quarter}{data quarter}

\item{year}{defaults to 2019}

\item{pivoted}{TRUE will turn EDTC data into long format}
}
\description{
Read and format EDTC files. Only used in the annual reports, not quarterly. And will likely need to change with 2020 annual report.
}
