% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hai_2019.R
\name{read_hai_2019}
\alias{read_hai_2019}
\title{Read HAI data}
\usage{
read_hai_2019(folder, quarter, year = 2020, joinMOU = FALSE)
}
\arguments{
\item{folder}{data location prior to quarter specification}

\item{quarter}{data quarter, used in path and elsewhere}

\item{year}{2020 by default}

\item{joinMOU}{whether the data should be joined the moulist file. needed for NA measures for non-respondents}
}
\description{
Read and combine HAI data files. Legacy: use read_hai_2020 as the default for this task. That reflects changes in data formatting. Use this function for running old reports.
}
