% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hcahps.R
\name{read_hcahps}
\alias{read_hcahps}
\title{Read HCAHPS data}
\usage{
read_hcahps(file, joinMOU = FALSE)
}
\arguments{
\item{file}{data location}

\item{joinMOU}{whether the data should be joined the moulist file. needed for NA measures for non-respondents}
}
\description{
Used for reading in HCAHPS data for quarterly reports
}
