% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rank.R
\name{reporting_rank}
\alias{reporting_rank}
\title{Rank States}
\usage{
reporting_rank(data, domain, report_year)
}
\arguments{
\item{data}{data file}

\item{domain}{either IP", "OP", "HCAHPS", or "EDTC"}

\item{report_year}{report year; removed default year}
}
\description{
This function is used to create the data for the side-by-side tables in the annual report.
}
