% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_plot.R
\name{res_plot}
\alias{res_plot}
\title{Scatterplot Matrix}
\usage{
res_plot(model)
}
\arguments{
\item{model}{model}
}
\description{
A quicker and better looking residual plot. takes the name of a model, for instance, "advert_model" from "advert_model <- lmSales ~ TV * Radio, data=advertising
}
