% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rndup1}
\alias{rndup1}
\title{Round data}
\usage{
rndup1(x, digits = 1)
}
\arguments{
\item{x}{data}

\item{digits}{desired number of digits, defaults to 1}
}
\description{
Round number to 1 digit, always rounds 0.5 up
}
