% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat_matrix.R
\name{scat_matrix}
\alias{scat_matrix}
\title{Scatterplot Matrix}
\usage{
scat_matrix(x, col = NULL)
}
\arguments{
\item{x}{data frame}

\item{col}{need to specify col as part of dataframe i.e. penguins$species-- this will provide some interesting feedback}
}
\description{
Function for a better looking/more functional scatterplot matrix than default pairs function and quicker than specifying with GGally
}
