% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vjust_hcahps2.R
\name{vjust_hcahps2}
\alias{vjust_hcahps2}
\title{Place Figure Text}
\usage{
vjust_hcahps2(
  data,
  axis_max = 100,
  axis_min = 0,
  invert = FALSE,
  upper = 0.925
)
}
\arguments{
\item{data}{data frame}

\item{axis_max}{maximum of fixed Y axis; default is 100; only used if fixed=T}

\item{axis_min}{minimum of fixed Y axis; default is 100; only used if fixed=T}

\item{invert}{use this option if plot shows the data in the opposite order as intended}

\item{upper}{the value past which all text gets placed below the line}
}
\description{
Used for positioning trend figure data value text appropriately in figures. This takes the state line into account for CAH figures.
}
