% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vjust_right.R
\name{vjust_right}
\alias{vjust_right}
\title{Place Figure Text}
\usage{
vjust_right(
  data,
  fixed = FALSE,
  axis_max = 100,
  axis_min = 0,
  invert = FALSE,
  upper = 0.9
)
}
\arguments{
\item{data}{data frame}

\item{fixed}{whether the plot's Y axis is fixed; default is FALSE}

\item{axis_max}{maximum of fixed Y axis; default is 100; only used if fixed=T}

\item{axis_min}{minimum of fixed Y axis; default is 100; only used if fixed=T}

\item{invert}{use this option if plot shows the data in the opposite order as intended}

\item{upper}{the percentage of the Y axis that will cut off data text if exceeded; defaults to .9}
}
\description{
Used for positioning trend figure data value text appropriately in figures
}
